feat:Refactor Node js with express router
