+++ backend/index.js
@@ -2,6 +2,7 @@ import express from 'express';
 import { PORT, mongoDBURL } from './config.js';
 import mongoose from 'mongoose';
 import { Book } from './models/bookModel.js';
+import booksRoute from './routes/booksRoute.js';
 
 const app = express();
 
@@ -13,97 +14,7 @@ app.get('/', (req, res) => {
   return res.status(234).send('Welcome To MERN Stack Tutorial');
 });
 
-// Route for Save a new Book
-app.post('/books', async (request, response) => {
-  try {
-    if (
-      !request.body.title ||
-      !request.body.author ||
-      !request.body.publishYear
-    ) {
-      return response.status(400).send({
-        message: 'Send all required fields: title,author,publishYear',
-      });
-    }
-    const newBook = {
-      title: request.body.title,
-      author: request.body.author,
-      publishYear: request.body.publishYear,
-    };
-
-    const book = await Book.create(newBook);
-    return response.status(201).send(book);
-  } catch (error) {
-    console.log(error.message);
-    response.status(500).send({ message: error.message });
-  }
-});
-
-// Route for Get All Book frm database
-app.get('/books', async (request, response) => {
-  try {
-    const books = await Book.find({});
-    return response.status(200).json({
-      count: books.length,
-      data: books,
-    });
-  } catch (error) {
-    console.log(error.message);
-    response.status(500).send({ message: error.message });
-  }
-});
-
-// Route for Get One Book frm database by id
-app.get('/books/:id', async (request, response) => {
-  try {
-    const { id } = request.params;
-    const book = await Book.findById(id);
-    return response.status(200).json(book);
-  } catch (error) {
-    console.log(error.message);
-    response.status(500).send({ message: error.message });
-  }
-});
-
-// Route for Update a Book
-app.put('/books/:id', async (request, response) => {
-  try {
-    if (
-      !request.body.title ||
-      !request.body.author ||
-      !request.body.publishYear
-    ) {
-      return response.status(400).send({
-        message: 'Send all required fields: title,author,publishYear',
-      });
-    }
-    const { id } = request.params;
-    const result = await Book.findByIdAndUpdate(id, request.body);
-
-    if (!result) {
-      return response.status(400).json({ message: 'Book not found' });
-    }
-    return response.status(200).send({ message: 'Book updated successfully' });
-  } catch (error) {
-    console.log(error.message);
-    response.status(500).send({ message: error.message });
-  }
-});
-
-//Route for Delete a Book
-app.delete('/books/:id', async (request, response) => {
-  try {
-    const { id } = request.params;
-    const result = await Book.findByIdAndDelete(id);
-    if (!result) {
-      return response.status(404).json({ message: 'Book not found' });
-    }
-    return response.status(200).send({ message: 'Book deleted successfully' });
-  } catch (error) {
-    console.log(error.message);
-    response.status(500).send({ message: error.message });
-  }
-});
+app.use('/books', booksRoute);
 
 mongoose
   .connect(mongoDBURL)
